chore(workflow): rename dir
